Added Enemy Attack Script
The Enemy Attack script is the script that will be attached to the box collider that the enemy spawns to attack the player. I referenced the Player and Enemy Animation scripts to the Enemy Attack script to check:

if the player is in the melee sphere collider then cue the attack animation and tell the Enemy Attack script to spawn the box collider.